requirements.txt
PyYAML==6.0.2
jsonschema==4.23.0